model-server/requirements.txt
flask==3.0.3
accelerate==0.33.0
transformers>=4.44.0
diffusers==0.29.2
torch>=2.0.0
torchvision
trimesh==4.4.9
numpy>=1.24.0
python-dotenv